# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED, INVALID_TOKEN, CONFLICT]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        origin_type:
          type: string
          nullable: true
          description: '"imported" when the listing is imported (else null) — drives the "مستورد / Imported" card badge on industrial + car listings.'
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
          description: Additive. The listing's main section — lets clients adapt per-section UI (e.g. the save glyph) without re-deriving it.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
        is_bookable:
          type: boolean
          nullable: true
          description: >-
            Additive. True when this listing is a furnished/daily rental (hotel
            model) that can be reserved from its detail. Lets the feed, search
            card, and map pin surface a "قابل للحجز / Bookable" affordance without
            fetching the detail. Null/false for everything else.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    DealerListing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        price_display:
          type: string
        price_raw:
          type: string
        location:
          type: string
        status:
          type: string
        created_at:
          type: string
        views:
          type: integer
        clicks:
          type: integer
        leads:
          type: integer
...
    SocialLink:
      type: object
      description: Public seller/company social link shown on the profile (Task #38).
      required: [platform, value]
      properties:
        platform:
          type: string
          enum: [instagram, linkedin, website, whatsapp]
        value:
          type: string
          description: URL (website/linkedin/instagram) or phone (whatsapp), normalized server-side.
    NotificationPreference:
      type: object
      description: Per-category notification preference. Absence of a row = enabled (Task #38).
      required: [type, in_app, email]
      properties:
        type:
          type: string
          enum: [message, lead, system, rfq, new_match, price_drop, comment, review, investment, global_supply, booking, payment_success, payment_failed, subscription_expiring]
        in_app:
          type: boolean
        email:
          type: boolean
    SavedSearch:
      type: object
      description: A saved search with optional price-tracking + match alerts (Task #38).
      required: [id, name, alerts_enabled, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        query:
          type: string
          nullable: true
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
        filters:
          type: object
          additionalProperties: true
          nullable: true
        price_min:
          type: string
          nullable: true
        price_max:
          type: string
          nullable: true
        alerts_enabled:
          type: boolean
        created_at:
          type: string
...
paths:
...
  /v1/me/listings:
    get:
      operationId: getMyListings
      summary: The caller's own listings (role-agnostic) for the profile grid
      tags: [me]
      security:
        - ClerkAuth: []
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated owner listings (FeedItem contract)
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/me/listings/manage:
    get:
      operationId: getMyManagedListings
      summary: The caller's own listings with management fields (role-agnostic)
      tags: [me]
      security:
        - ClerkAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, sold, archived]
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at, price, views, leads]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated owner listings with management fields
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DealerListing"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/me/social-links:
    get:
      operationId: getMySocialLinks
      summary: Get my profile social links
      tags: [me]
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: My social links
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SocialLink"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    put:
      operationId: setMySocialLinks
      summary: Replace my full set of social links
      tags: [me]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [links]
              properties:
                links:
                  type: array
                  items:
                    $ref: "#/components/schemas/SocialLink"
      responses:
        "200":
          description: Updated social links
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SocialLink"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/me/notification-preferences:
    get:
      operationId: getMyNotificationPreferences
      summary: Get my per-category notification preferences
      tags: [me]
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: My notification preferences (defaults applied for missing categories)
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationPreference"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    put:
      operationId: setMyNotificationPreferences
      summary: Update my notification preferences (mute/unmute categories)
      tags: [me]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preferences]
              properties:
                preferences:
                  type: array
                  items:
                    $ref: "#/components/schemas/NotificationPreference"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationPreference"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/me/saved-searches:
    get:
      operationId: listMySavedSearches
      summary: List my saved searches
      tags: [me]
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: My saved searches (newest first)
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SavedSearch"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    post:
      operationId: createSavedSearch
      summary: Save a search (with optional price tracking + alerts)
      tags: [me]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                query:
                  type: string
                  nullable: true
                category:
                  type: string
                  enum: [car, real_estate, industrial]
                  nullable: true
                filters:
                  type: object
                  additionalProperties: true
                  nullable: true
                price_min:
                  type: string
                  nullable: true
                price_max:
                  type: string
                  nullable: true
                alerts_enabled:
                  type: boolean
      responses:
        "200":
          description: The created saved search
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/SavedSearch"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/me/saved-searches/{id}:
    patch:
      operationId: updateSavedSearch
      summary: Update a saved search (rename, toggle alerts, adjust price range)
      tags: [me]
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price_min:
                  type: string
                  nullable: true
                price_max:
                  type: string
                  nullable: true
                alerts_enabled:
                  type: boolean
      responses:
        "200":
          description: The updated saved search
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    $ref: "#/components/schemas/SavedSearch"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    delete:
      operationId: deleteSavedSearch
      summary: Delete a saved search
      tags: [me]
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                required: [data, error, meta]
                properties:
                  data:
                    type: object
                    required: [deleted]
                    properties:
                      deleted:
                        type: boolean
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...